+++ src/Sentry/Infrastructure/DebugDiagnosticLogger.cs
@@ -1,6 +1,5 @@
 using System;
 using System.Diagnostics;
-using Sentry.Extensibility;
 
 namespace Sentry.Infrastructure
 {
@@ -11,25 +10,16 @@ namespace Sentry.Infrastructure
     /// Logger available when compiled in Debug mode. It's useful when debugging apps running under IIS which have no output to Console logger.
     /// </remarks>
     [Obsolete("Logger doesn't work outside of Sentry SDK. Please use TraceDiagnosticLogger instead")]
-    public class DebugDiagnosticLogger : IDiagnosticLogger
+    public class DebugDiagnosticLogger : DiagnosticLogger
     {
-        private readonly SentryLevel _minimalLevel;
-
         /// <summary>
         /// Creates a new instance of <see cref="DebugDiagnosticLogger"/>.
         /// </summary>
-        public DebugDiagnosticLogger(SentryLevel minimalLevel) => _minimalLevel = minimalLevel;
+        public DebugDiagnosticLogger(SentryLevel minimalLevel) : base(minimalLevel)
+        {
+        }
 
-        /// <summary>
-        /// Whether the logger is enabled to the defined level.
-        /// </summary>
-        public bool IsEnabled(SentryLevel level) => level >= _minimalLevel;
-
-        /// <summary>
-        /// Log message with level, exception and parameters.
-        /// </summary>
-        public void Log(SentryLevel logLevel, string message, Exception? exception = null, params object?[] args)
-            => Debug.Write($@"{logLevel,7}: {string.Format(message, args)}
-{exception}");
+        /// <inheritdoc />
+        protected override void LogMessage(string message) => Debug.WriteLine(message);
     }
 }